requirements.txt
amazon-textract-textractor==1.8.2
colorama==0.4.6
docx2pdf==0.1.8
docx2txt==0.8
effdet==0.4.1
email-validator==2.2.0
flask-cors==4.0.1
fqdn==1.5.1
google-cloud-vision==3.7.2
grpcio-status==1.62.2
httptools==0.6.1
isoduration==20.11.0
langchain-chroma==0.1.2
langchain-cli==0.0.25
langchain-community==0.2.5
langchain-groq==0.1.6
langchain-openai==0.1.14
langchain-pinecone==0.1.1
langchainhub==0.1.20
lark==1.1.9
markdown==3.6
notebook==7.2.1
openpyxl==3.1.4
pikepdf==9.0.0
pillow-heif==0.16.0
pymupdf==1.24.7
pytesseract==0.3.10
python-dotenv==1.0.1
python-pptx==0.6.23
rapidocr-onnxruntime==1.3.22
rich==13.7.1
sentence-transformers==3.0.1
shellingham==1.5.4
sse-starlette==1.8.2
ujson==5.10.0
unstructured==0.14.8
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
uri-template==1.3.0
uvloop==0.19.0
watchfiles==0.22.0
webcolors==24.6.0
websockets==12.0

-e .